Fine tuning of export functions and dialogs.
