+++ satkit/export/__init__.py
@@ -33,11 +33,12 @@
 
 from .images import (
     export_aggregate_image_and_meta,
-    export_aggregate_image_meta,
     export_distance_matrix_and_meta,
     export_ultrasound_frame_and_meta,
-    export_ultrasound_frame_meta,
     publish_distance_matrix,
     publish_aggregate_images
 )
-
+from .meta_data import (
+    export_aggregate_image_meta, export_session_and_recording_meta,
+    export_ultrasound_frame_meta
+)
